+++ src/test/java/fopbot_playground/PacmanRobot1Example.java
@@ -1,4 +1,4 @@
-package fopbot_playground.example;
+package fopbot_playground;
 
 import fopbot.World;
